function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
psycopg2-binary
python-http-client==3.3.7
sendgrid==6.11.0
starkbank-ecdsa==2.2.0